yaft: fractional font scaling (1.5x support)
Changed fontScale from int to float. The glyph drawing loop now uses
output-driven block sizes (int((w+1)*fs) - int(w*fs)) so fractional scales
like 1.5 produce correct pixel blocks without gaps or overlaps.

+++ apps/yaft/config.h
@@ -38,9 +38,9 @@ struct YaftConfig {
   // Key repeat rate in chars / sec.
   int repeatRate = 25;
 
-  // Font scale: 1 = default (16x32 cells), 2 = double size (32x64).
+  // Font scale: 1.0 = default (16x32 cells), 1.5 = 50% bigger, 2.0 = double.
   // Scales the BDF glyph rendering so text is bigger (and fewer cells fit).
-  int fontScale = 1;
+  float fontScale = 1.0f;
 
   static YaftConfig getDefault();
 };
